this is my favorite language ever

+++ ProjectEuler/PE14/script.jl
@@ -1,4 +1,4 @@
-collatz(n::Int, i=0) = n == 1 ? i : n % 2 == 0 ? collatz(n/2, i+1) : collatz((3n+1)/2, i+2)
+collatz(n, i=0) = n == 1 ? i : n % 2 == 0 ? collatz(n/2, i+1) : collatz((3n+1)/2, i+2)
 
 function main()
     collatz.([1:10^6;]) |> findmax
